requirements.txt
discord.py
criptocompare
configparser